Refactor rendering of Form Node and Form Completion Node
